Fix test duplication.
