+++ packages/gensx/tests/checkpoint.test.tsx
@@ -1,71 +1,21 @@
 import { setTimeout } from "timers/promises";
 
 import type { ExecutionNode } from "@/checkpoint.js";
-import type { ExecutableValue } from "@/types.js";
 
 import { afterEach, beforeEach, expect, suite, test, vi } from "vitest";
 
 import { CheckpointManager } from "@/checkpoint.js";
-import { ExecutionContext, withContext } from "@/context.js";
 import { gsx } from "@/index.js";
-import { createWorkflowContext } from "@/workflow-context.js";
-
-// Add types for fetch API
-export type FetchInput = Parameters<typeof fetch>[0];
-export type FetchInit = Parameters<typeof fetch>[1];
+import {
+  executeWithCheckpoints,
+  FetchInit,
+  FetchInput,
+  getExecutionFromBody,
+} from "./utils/executeWithCheckpoints";
 
 // Helper function to generate test IDs
 export function generateTestId(): string {
   return `test-${Math.random().toString(36).substring(7)}`;
-}
-
-/**
- * Helper to execute a workflow with checkpoint tracking
- * Returns both the execution result and recorded checkpoints for verification
- */
-
-// eslint-disable-next-line @typescript-eslint/no-unnecessary-type-parameters
-export async function executeWithCheckpoints<T>(
-  element: ExecutableValue,
-): Promise<{
-  result: T;
-  checkpoints: ExecutionNode[];
-  checkpointManager: CheckpointManager;
-}> {
-  const checkpoints: ExecutionNode[] = [];
-
-  // Set up fetch mock to capture checkpoints
-  global.fetch = vi
-    .fn()
-    // eslint-disable-next-line @typescript-eslint/require-await
-    .mockImplementation(async (_input: FetchInput, options?: FetchInit) => {
-      if (!options?.body) throw new Error("No body provided");
-      const checkpoint = JSON.parse(
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-unsafe-member-access
-        JSON.parse(options.body as string).rawExecution,
-      ) as ExecutionNode;
-      checkpoints.push(checkpoint);
-      return new Response(null, { status: 200 });
-    });
-
-  // Create and configure workflow context
-  const checkpointManager = new CheckpointManager();
-  const workflowContext = createWorkflowContext();
-  workflowContext.checkpointManager = checkpointManager;
-  const executionContext = new ExecutionContext({});
-  const contextWithWorkflow = executionContext.withContext({
-    [Symbol.for("gensx.workflow")]: workflowContext,
-  });
-
-  // Execute with context
-  const result = await withContext(contextWithWorkflow, () =>
-    gsx.execute<T>(element),
-  );
-
-  // Wait for any pending checkpoints
-  await checkpointManager.waitForPendingUpdates();
-
-  return { result, checkpoints, checkpointManager };
 }
 
 suite("checkpoint", () => {
@@ -538,10 +488,7 @@ suite("tree reconstruction", () => {
     expect(lastCall).toBeDefined();
     const options = lastCall![1] as FetchInit;
     expect(options?.body).toBeDefined();
-    const lastCallBody = JSON.parse(
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-unsafe-member-access
-      JSON.parse(options!.body! as string).rawExecution,
-    ) as ExecutionNode;
+    const lastCallBody = getExecutionFromBody(options?.body as string);
     expect(lastCallBody.componentName).toBe("Parent");
     expect(lastCallBody.children[0].componentName).toBe("Child1");
